gitlab-ci/tools/codeowners_tools/pyproject.toml
[tool.poetry]
name = "codeowners_tools"
version = "0.1.0"
description = ""
authors = ["Valeriy Zamaraiev <valeriy.zamaraiev@dfinity.org>"]
include = ["user_map.csv"]

[tool.poetry.dependencies]
python = "^3.9"
python-gitlab = "^2.10.0"
GitPython = "^3.1.20"
PyGithub = "^1.55"
absl-py = "^0.13.0"
codeowners = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"


[tool.poetry.scripts]
codeowners_check = "codeowners_tools.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
